requirements.txt
Django~=5.0
djangorestframework~=3.15
djangorestframework-simplejwt~=5.3
psycopg[binary]~=3.1
python-decouple~=3.8
